Fix type move
