+++ packages/ag-grid-community/src/headerRendering/cells/column/headerCellComp.ts
@@ -1,4 +1,5 @@
 import type { HeaderStyle } from '../../../entities/colDef';
+import type { IHeaderComp } from '../../../interfaces/iHeader';
 import type { UserCompDetails } from '../../../interfaces/iUserCompDetails';
 import { _removeAriaSort, _setAriaSort } from '../../../utils/aria';
 import type { ElementParams } from '../../../utils/dom';
@@ -6,7 +7,6 @@ import { _addStylesToElement } from '../../../utils/dom';
 import { RefPlaceholder } from '../../../widgets/component';
 import { AbstractHeaderCellComp } from '../abstractCell/abstractHeaderCellComp';
 import type { HeaderCellCtrl, IHeaderCellComp } from './headerCellCtrl';
-import type { IHeaderComp } from './headerComp';
 
 const HeaderCellElement: ElementParams = {
     tag: 'div',
